refactor(planner): refine the file organization
