+++ src/query/service/src/interpreters/interpreter_table_recluster.rs
@@ -54,10 +54,10 @@ use databend_common_sql::plans::set_update_stream_columns;
 use databend_common_sql::plans::BoundColumnRef;
 use databend_common_sql::plans::Plan;
 use databend_common_sql::plans::ReclusterPlan;
-use databend_common_sql::query_executor::QueryExecutor;
 use databend_common_sql::IdentifierNormalizer;
 use databend_common_sql::MetadataRef;
 use databend_common_sql::NameResolutionContext;
+use databend_common_sql::QueryExecutor;
 use databend_common_sql::ScalarExpr;
 use databend_common_sql::TypeChecker;
 use databend_enterprise_hilbert_clustering::get_hilbert_clustering_handler;
