+++ src/query/sql/src/planner/execution/mod.rs
@@ -1,0 +1,21 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+mod query_executor;
+mod stream_column;
+
+pub use query_executor::QueryExecutor;
+pub use stream_column::StreamContext;
+pub use stream_column::CURRENT_BLOCK_ID_COL_NAME;
+pub use stream_column::CURRENT_BLOCK_ROW_NUM_COL_NAME;
