+++ src/query/sql/src/planner/execution/stream_column.rs
