+++ src/query/sql/src/planner/optimizer/dynamic_sample/dynamic_sample.rs
@@ -25,7 +25,7 @@ use crate::optimizer::dynamic_sample::join_selectivity_sample::join_selectivity_
 use crate::optimizer::ir::RelExpr;
 use crate::optimizer::ir::SExpr;
 use crate::optimizer::ir::StatInfo;
-use crate::planner::query_executor::QueryExecutor;
+use crate::planner::QueryExecutor;
 use crate::plans::Aggregate;
 use crate::plans::AggregateMode;
 use crate::plans::Limit;
