+++ src/query/sql/src/planner/optimizer/hyper_dp/dphyp.rs
@@ -32,7 +32,7 @@ use crate::optimizer::rule::TransformResult;
 use crate::optimizer::OptimizerContext;
 use crate::optimizer::RuleFactory;
 use crate::optimizer::RuleID;
-use crate::planner::query_executor::QueryExecutor;
+use crate::planner::QueryExecutor;
 use crate::plans::Filter;
 use crate::plans::JoinType;
 use crate::plans::RelOperator;
