+++ src/query/sql/src/planner/semantic/type_check.rs
@@ -126,9 +126,9 @@ use crate::binder::NameResolutionResult;
 use crate::optimizer::ir::RelExpr;
 use crate::optimizer::ir::SExpr;
 use crate::parse_lambda_expr;
+use crate::planner::expression::UDFValidator;
 use crate::planner::metadata::optimize_remove_count_args;
 use crate::planner::semantic::lowering::TypeCheck;
-use crate::planner::udf_validator::UDFValidator;
 use crate::plans::Aggregate;
 use crate::plans::AggregateFunction;
 use crate::plans::AggregateFunctionScalarSortDesc;
